Add files via upload
Following previous studies, the initial opinion of each individual is generated randomly, where all approaches have the same initial opinion set. It should be noted that the final results will be greatly influenced by the initial opinion set. In this case, we repeatedly conduct every experiment 20 times (with different initial opinion sets), and report the average results of the 20 experiments for each approach.